Fix endless image errors
With the current setup, `dist/images/generic.png` is not available in my
React app. I tried prepending the `node_modules` path, which also
failed.

It fails rather spectacularly, quickly generating tens of thousands of
failed network requests and console errors, eventually crashing the tab.
I think it is the issue referenced here:
https://github.com/almende/vis/issues/3338

I found that the 'dot' shape is a sane default.